Update randomizer.py
(1). Added hints. 4 NPCs will now give you information about items and locations. They are...

The left NPC in Alsedo after beating Myconid.
The left and right NPCs at the top of the stairs in Purapil.
The "Amulet Lady". 

(2). Legend Sword added to the pool. You must find it to beat the game.

(3). Ocarina fixed and now in pool. 

(4). Various speed ups added including playing ocarina, talking to the sphinx, and the opening credits.